refactor: manager 네이밍 변경
