+++ src/js/videoAPI.js
@@ -1,6 +1,6 @@
-import { ERROR_MESSAGE, VIDEO_LIST } from '../utils/constants.js';
+import { ERROR_MESSAGE, VIDEO_LIST } from './utils/constants.js';
 
-const APIManager = {
+const videoAPI = {
   baseURL: 'https://hardcore-williams-b69905.netlify.app/youtube/v3/search',
   pageToken: '',
   part: 'snippet',
@@ -49,4 +49,4 @@ const APIManager = {
   },
 };
 
-export default APIManager;
+export default videoAPI;
